requirements.txt
pymysql >= 0.10.1
sqlmodel
sqlalchemy
sqlacodegen
mysql-connector-python
pillow
cryptography